src/server/requirements.txt
Flask>=1.*
SQLAlchemy>=1.3.6
mysql-connector>=2.*
requests==2.2.1
pyyaml>=3.13
Coverage>=4.5
requests-mock>=1.5.2
Faker>=1.0.4
testing.mysqld>=1.4.*